refactor(math): fold geometry domains under geometry
Fold the euclidean, exact, and projective geometry operation domains in as
subpackages of the surviving geometry owner. Operation IDs and schemas are
unchanged.

- euclidean_geometry -> geometry/euclidean
- exact_geometry -> geometry/exact
- projective_geometry -> geometry/projective

Update the one cross-family import (polynomial syzygy models consume
RationalProjectiveLine) and the import-linter source module.

Validation: make check green (ruff, mypy, 977 tests); import-contracts
and architecture pass; snapshot unchanged at 360.
