+++ src/jacobian/math/geometry/euclidean/__init__.py
