+++ src/jacobian/math/geometry/euclidean/_models.py
