+++ src/jacobian/math/geometry/exact/_models.py
